requirements.txt
goatools
biopython
requests
pronto